requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
python-decouple==3.8
pydantic==1.10.9
email-validator==2.0.0
dnspython==2.3.0
SQLAlchemy==2.0.16
python-jose==3.3.0
passlib==1.7.4
#psycopg2-binary==2.9.6
emails==0.6
alembic==1.11.1
asyncpg==0.27.0
fastapi-users[sqlalchemy]
fastapi-users-db-sqlalchemy==5.0.0